agents/legal_researcher/pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "legal-researcher-agent"
version = "1.0.0"
description = "Legal Researcher Agent - RAG for Legal Precedents"
requires-python = ">=3.10"

dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-aiplatform>=1.40.0",
    "google-genai>=0.3.0",
]
